Added a single entity that moves around the world.
Imported the entity system, and created three components: A transform
component (which currently only consists of a transformation matrix), a
'rail denizen' component which is a thing that follows a rail, and a
player component, which is currently empty but will represent any player
character in the world.

Additionally, the game class loads the config file and takes the first
component it finds and uses it's rail denizen component to drive the
camera. This will change in the future.

Functions to draw white boxes have also been added so that we have some
dummy scenery for development.

Note: There is an issue with motivators that causes the game to crash
after the camera has moved a certain distance. A fix for this will be
checked in shortly, but I wanted to check in the work that I've done so
far so that I'm not blocking others.

Change-Id: Ie8032c35741215911105ecbf8fb174e388fed0c4