Added basic physics component
Added in a basic physics component.  It doesn't do much at the
moment, but it will be useful soon as an endpoint that can be
replaced with an actual interface to bullet, once we get that in.

Also changed the transforms over to using position, scale, and
a quaternion for orientation, instead of just a raw vector.

Change-Id: Ifc33ab098beb64439fe0047ab92401617df7466f