Update build_assets.py to also build meshes.
Added a simple cube mesh, to show the pipeline is working.
We'll eventually delete the cube mesh.

The pipeline requires that you manually add material.json
files. For example, we have to manually add cube__guy.json,
for the texture cube__guy.png, which gets pulled out of
the source model cube.fbx. It would be nicer if the
pipeline automatically determined the material format and
created cube__guy.json, but for now this is what we have.

Tested: Works on Linux.
Change-Id: I55004806a0ea507e7b79527af4718360cbfa1f7c