Load cube mesh from cube.fplmesh file.
Delete the function that creates the cube mesh procedurally.

The only difference between them is that cube.fplmesh doesn't
have tangents (though it has normals), and cube.fplmesh is
centered at 0 (for z--up and down).

We increase the height of the boxes on screen by 1 so that
everything renders the same.

Tested: Runs looking the same on Linux.
Change-Id: I5e869a78827366212cde20c813027e1c0bfab2de