Added some sounds to projectiles.
Added code to play sounds when projetiles are launched or when they hit
the ground, which meant piping the audio engine around to the places
that need it to play sounds.

Added some ogg files (created by me, so they have no license
restrictions) and the associated json files that define the sounds'
attributes.

Bug: 20175092

Change-Id: I90913f993e8e7b60e6c4033613d771bf5bf21a86