Fixed memory corruption due to IntrusiveLists.
The copy and move constructors were not set up correctly, causing them
to point at freed memory when copied around in, for example,
std::vectors. I've added a move constructor and disallowed copying
because copying an intrusive list doesn't really make sense.

Also removed the TypedIntrusiveList version of the class since we're not
using it, and I'd like to simplify things.

Tested on Linux. Can't test on windows because of a bunch of other
breakages, but my reasearch suggests this will work on the versions of
MSVC we support.

Bug: 21084955
Change-Id: Id9a708104f360c6ac4ff8f3b0df9d1d375b77475