Allow components to export raw data to be later imported.
This adds the ExportRawData function to the component interface. Any
component which wants to be able to serialize its state (for example,
for use in an editor) should implement this function to export its
data. It can later re-import the data via the existing AddFromRawData
function.

This also includes a sample implementation for TransformData.

Change-Id: If76e30eed3c044fe439224a75509a0e87771d3db