Added an event system.
An event manager has been added that can route events sent from on
location to any listener that cares about them. Events may also have a
payload associated with them. Payloads are declared much like entities
and components are: with a macro that associates a unique event id with
a given payload type.

Event listeners may register themselves with an event manager and all
events will be routed to the listener's OnEvent function.

Currently I've set up the GameState as both the listener and the owner
of the EventManager, but this can change in the future if we find a
better organizational method.

An event ID for playing sounds has been added and hooked up to the
physics component.

Change-Id: I0dae814a1a148b8afd6d436246e03e2a2479016b