Added Sound and Listener components.
A sound component is for playing any continuous sound at an entity's
position (such as the whooshing sound made from throwing sushi. Also for
things like campfires).

A listener component is determines where the audio engine listener is
going to be. Like the sound component, the listener is set to the
entity's position.

Also updated the data to set the listener on player entity and sounds on
projectiles.

Change-Id: I1721bba32b2ba27685ff2754fa7dfff344fcb6a1