Added a score component that tracks players scores.
Currently the scores are tracked but nothing is done with them.

Patrons now broadcast events when hit: When a patron is hit, they
broadcast an event announcing they've been hit. They also broadcast an
event that states whether they were hit in the mouth or on their body.
Additionally, an event is broadcast when a projectile is fired.

The score component has been added to the game state and is hooked up to
the event manager.

Tested on Linux.
Bug: 20821045

Change-Id: Ib7705219940ae92fce6835c27328ad311456066c