Added UI elements for in-game scorekeeping.
This adds a very simple and rudimentry score UI. For each entity with a
score component, a score counter will be displayed in the corner of the
screen. This is a preliminary UI right now, and we can improve it when
ImGui is more fleshed out and when I can figure out how to position
things in a better way.

Tested on Linux

Change-Id: I6401883d1e6cac6686d20f9eab36a455c3ef292d