Added simple cone-frustrum culling to the rendermesh component.
The rendermesh component will no longer render objects that are
outside the cone representing the field of view.

Also made some minor tweaks to the world definition so that some
of the larger objects in the world were a little smaller.  This is
to minimize the amount of fudge-factor needed to make sure that
things that are just outside of the view cone wouldn't pop-in at
the edge of the screen.

Change-Id: I34590959ce74d293f78a0bdba35aa75f49f9d405