Add Lady Mandrill and Menacing Croc.
Add textures and meshes for the above patrons.

Also add EntityDefs to differentiate between the different
patrons. Each patron shows up in the config so they can be
seen in the game. For now their hit boxes are reused from
the Moustache Croc.

Renamed `croc_moustache` to `moustache_croc`.

Trimmed texture coordinates and geometry to avoid big bars
of alpha on either side of the billboards.

Tested: Game runs on Mac.
Change-Id: Ib010c7eb5dc602ee8bc41f51360755bff62667ed