Fixed assertion in score when firing events.
The issue was caused by sending a structure which was not registered
with the event registry by mistake. This was legal because the function
was templated and could accept any type, and an incorrect type was being
passed to it. To prevent this from happening again, a static_assert has
been added to prevent using unregistered types.

Tested on Linux.

Change-Id: Ia97e479a288f3f0008a2e72c7fee192399ef6e60