Added entity prototyping system & basic transform editing.
All entities in the game can now be loaded from prototypes in
entity_library.json.

You can also do rudimentary transform editing of entities.

Use [ ] to cycle between objects in the game. When an object is
selected, use IJKLP; to move it, and UOYHNM to rotate it. Use +-
to scale it.

Press INS to copy the selected entity, and DEL to delete it.

Press F11 while in edit mode to save all entities to their
respective lists.

Also supports clicking the mouse to select an entity using
ray casting.

Tested: Desktop build tested on Linux and Mac. Android built
on Linux and Mac but not tested.

Change-Id: I9270baba16c1616d2a29e790cb8d99e9d59c465e