RegisterComponent no longer requires type.
Previously, calls to RegisterComponent required that you add the type of
the argument, like so:

  entity_manager_.RegisterComponent<TransformComponent>(&transform_component_);

This changes RegisterComponent to take an argument of type T instead of
type ComponentInterface, which means the argument type can be derived by
the call. It's also still required to inherit from ComponentInterface
because the helper function that's called immediately has a
ComponentInterface* as it's argument.

This also means you can't accidentally introduce a mismatch between a
component object and its type when calling the function (e.g. you can no
longer call `RegisterComponent<FooComponent>(bar_component)` by mistake).

Change-Id: I09729ff94b99578c1ef3f902b15b7d69b313cfa8