Events can now be specified in data.
This is a pretty large change to how events are piped around the game.
Rather than have events that represent certain kinds of events that
happened in game (like patrons getting hit), there are events for
different kinds of actions you want to take.

To that end, PatronHit events have been removed and been replaced with
AttributeModification events. Also a ChangeRailSpeed event has been
added, but it is nonfunctional until Motive supports applying a scale to
a spline.

Added an ActionDef schema which represents an action that can be taken
in response to an event. These ActionDefs have been hooked up in various
locations in the config file, such as on patrons and physics objects so
that they can be broadcast when appropriate.

Added a ParseAction function that can turn the ActionDef into an actual
action object given the current context and passes it to the event
manager to be handled.

Also changed the name of the score component to attribute component.

Tested on Linux and Android.

Change-Id: I47d80391e82f16274b19eede3572fb0f16765529