Adds Collision handling through Bullet physics
Uses Bullet to generate collision events, which are in used by patron
to determine when one has been fed.  Attaches simple billboard physics
definitions to the various patrons.

Adds a way to specify an offset of the collision shape from the main
transform, along with a way to mark a body as kinematic.

Change-Id: I9ec85f00dee3c51da4953a24eebada0e9ec3f104