Patrons will now only stand up when in range.
Added 'pop in' and 'pop out' values to patrons, which control when they
stand up and when they fall back down. If the distance between you and
the patron is less than the pop in radius, the patron will stand up. If
the distance between you and the patron is greater than the pop out
radius (which is must be greater than or equal to the pop in radius) then
they will fall back down.

Tested on Linux
Bug: 21640564

Change-Id: I90170aaac0a484c4514399724b6a7fe800f802b4