Setting up a level layout based on the design docs
Extending the rail to take a minute to go along, and laying out
patrons that get fed across the multiple laps. Reduces the scale of
a lot of objects, bringing them closer to realistic in terms of having
a unit equal a meter.

Also fixes up some patron physics disable/enable timing issues.

Change-Id: Ie3a7f8a7931d3a844af555206b6236e9727a1bb1