Refactored the GameState class.
Added a new StateMachine class for handling the various states the game
can be in. Currently there are just two: Gameplay and editor. Eventually
more will be added. This new organization scheme is not perfect, but it
is an improvement over what we have now.

The GameState class has been split into three seperate classes: World,
which represents all the entities and the data that's associated with
their behavior and appearance; GameplayState, which has a pointer to a
world and handles updating the game each frame; and WorldEditorState,
which handles editing the world.

The WorldEditor code no longer needed an is_active_ boolean, so that was
removed.

Tested on Linux and Android.

Change-Id: I78e8eb346a2935ce7fe629a56ee2d47b34f7543e