Made the camera relative to the raft.
The raft is now a separate entity from the player, and their
orientations are now property composed. This turned out to be a little
bit bigger of a change than I intended because a few other things were
expecting the raft and the player to be the same entity, and also
because the projectiles could no longer rely on the player's orientation
and position to be in world space, so helper functions were written to
transform an entity's local position (relative to their parent) into
world space.

Tested on Linux and Android.

Change-Id: I2006b7ade152c166a3b5149c4db053941e22b327