Added a digit component, raft and sign.
The digit component that updates an entity's mesh based on some
attribute of the player. This is used for displaying meshes in the world
that represents score on signs.

Also hooked up the raft mesh and threw a sign on it that utilizes the
digit component to display your score.

Tested on Linux and Android.

Change-Id: I3ab57654c817d423bdd50a72e269dfc54e0c2cac