Added render & selection options to entities in-editor.
You can specify a selection mode for entities:
Any allows the entity to be selected by the pointer or by
cycling through entities.

PointerOnly only allows you to select the entity by clicking
on it.

CycleOnly only allows you to cycle through it, not click
on it.

None makes the entity not selectable in the editor.

You can also specify a special rendering mode for entities in
the editor. You can specify them to only render inside the editor
or only render outside the editor, or render always (the default).

Change-Id: I4ab5ba1f09908fc126023169aacb2d0b901b5ee5
Tested: Sample selection options are in the existing asset json files.