Moved rail definitions to components, and made editable.
Rails are now defined in entity_rails.json, as a series of entities
with a RailNode component definition.

RailNode specifies the order of the nodes, and the rail name
(a string); RailManager can then get you a rail from the strings.

Tested: Press F10 to go into the editor, then select a rail node
and move it with IJKL. After you exit, you will follow the new rail.

Change-Id: I46bbe4b25914ea56ea46c18ebc5a73ce14bfaf4d