Decoupled entity ComponentID from FlatBuffers data type.
Component ID is no longer statically defined when creating the
source file. Instead, it's set by the entity manager based on the
order in which the components are registered.

This is the first step in making a standard component library.

Change-Id: I1c9a88efa8b2d8943229140ff7f0035a8cea5d64
Tested: With this change and http://ag/722540, the game still builds
and runs properly, including the editor.