Added a tiny chime any time to score a point.
Added a chime ogg file. I generated the file myself, I didn't grab it
from any license encumbered source. A sound collection file was added
and declared in the sound bank, and that new sound collection is
referenced in a PlaySound event that occurs when points are scored.

Tested on Linux.

Change-Id: I41424726a7aaf5215dfc64784ca8f627004cb0e0