Fix bank flickering whenever a river node gets moved.
The bank was getting regenerated, and part of the regeneration
involves randomness. This change uses a fixed random seed to
calculate the same values every time.

The seed is fixed at the asset level, so the banks will be
generated the same every time the game is run.

Tested: Works on Linux.
Change-Id: I92b1f6d3f414a29451a7ff34a129acb7aad84aef