Move the world editor into its own library.
The editor is now part of the world_editor library, which uses
standard components that have previously been moved to the
component_library subdirectory in the entity library.

Depends on I87801247, I6d2125bc, and I6e728a80.

Tested: Built and ran on Mac desktop. Builds on Linux and for Android.
Change-Id: Id7929bc6142b0060ab3b36f17e1ca57bcc1ac967