Added a way to pause and return to the title.
Added a function to reset the world state. This is essentially the
world's initialize function split off into two functions: one to do core
initialization that we can keep between loading different levels, and
one that just loads the level which can be called multiple times.

Added a pause menu state, and changed the game state transitions so that
hitting back or escape brings you to the pause menu during the game (and
exits the game from the main menu).

Cleaned up the Components flatbuffer schema by removing the old
deprecated fields. Also added a WorldDef that contains all the data
needed for setting up the game world. It's currently just a list of file
names to load entity definitions from.

Tested on Linux.
Bug: 22451293

Change-Id: Ie85e9f75c3c08f08246ff0ddc95fe933e39c60b7