Specify texture format via pipeline metadata file.
New metadata file asset_meta.json is parsed by
build_assets.py. The metadata is applied to assets
that are processed.

The current change adds metadata only for the mesh fbx
files. You can specify the target texture formats.
We need this for the skybox fbxs, that must be in 888
format, instead of the default 565.

Change-Id: I34730711304a13e57710123d6bc2fe6c2fad6018
Tested: Works on Linux.