Added a simple light shader for normal-based lighting.
Adds a shader to handle simple lighting for objects that
never have shadows cast on them.  Also converted most objects
in the game over to use it.  (The only objects that receives
shadows now is the riverbank.)

Also updated most of the json files for geometry, to include
normals facing in both view directions.

Caused about a 15-20% speed increase on Nexus 5.

Tested on Linux and Android.

Change-Id: I78b57fd8fc293f9bb80ee8490a7e465bcd5aff22