Added new sounds and updated physics to play them.
Added the new sound effects and ambiance and removed most of the old
sound effects.

Changed the Patron so that they can react to multiple different
collision events. Events are differentiated by an arbitrary tag placed
on them when they occur. Rather than have a single on_collide ActionDef,
they have a list of possible on_collision actions that are tagged, and
only the matching action is executed. e.g. Hitting the mouth will create
a collision tagged with "Mouth", which will then execute the "Mouth"
ActionDef. In the long term, we want to move this on_collision behavior
to the physics component itself, but that currently requires hoisting a
bunch of code out of Zooshi and into the Event system, so I'll do that
in a seperate checkin. b/22847175

Hooked up the audio engine to a few states so that sounds could play on
state transitions. This would be better handled through the event
system, but that plumbing isn't hooked up yet. b/22851325

Tested on Linux and Android.

Change-Id: Id6925376e635def0d131d9f4e939e4ac034a9579