Fixing up the scale on the sushi projectile
The physics component is now taking into account the scale of the
transform component, so the sushi prototype needs to have its physics
shape reduced.

Also fixes a bug where child entities were not being instantiated on
thrown sushi.

Tested on Linux, using the editor to confirm the sizes are correct.

Change-Id: I3a98011ba3866c27f2829f03e0fbddeec598ae1a