Fixed entity positions passed to audio engine.
Fixed two issues with the audio: First, the position of audio samples
was being set to the entity's location is local space (i.e. it's
location relative to its parent) rather than in world space, meaning
that a lot of sounds were being played at 0,0,0.

Second, fixed the orientation of the listener. For some reason,
previously we were setting the position but not orientation of the
listener. Listeners now propery take the entity's entire matrix into
account.

Tested on Linux and OSX.

Change-Id: Ib5347d25b666abfd30e9d981061b340dc199d622