Added music and remaining animal sound effects.
Hooked up different audio for the game and menu states. The way this is
hooked up is temporary as doing it right requires the event system.
Eventually the gameplay music will only kick in once you feed the first
animal.

Also hooked up the chomp sound effects for the remaining animals.

Tested on Linux.

Change-Id: Id75f3f66de3ed895eb557cee39260aa86f0639fe