Tracking the current lap on a rail with a float.
Changing the rail denizen component to use a float when tracking the
current lap, instead of just an int.  This will allow behavior to fire
at different points of the laps, such as the soon to be added cave
at the beginning.

Note that the current use of the lap number, the patrons, have also
been updated to use a min and max lap of a float, allowing more
precise control of when they appear.

Tested on Linux.

Change-Id: I4716b3c19b7e908507d76f643584c7b13fbe84af