Changed to Image/GraphicsMagick from PIL in build_assets.py.
Changed to ImageMagick or GraphicsMagick since PIL installation can
potentially be tricky.
This means users only need to download and add ImageMagick or GraphicsMagick
tools to their path to build Zooshi.

ImageMagick can be slow so the build script which select GraphicsMagick if
it's available.

Tested:
Verified assets are built successfully on Linux.

Bug: 22628302

Change-Id: Ieece22ada73c76af5a5bd09faa2847110ce8d97c