Patrons move to catch sushi.
Position and orientation are adjusted to catch sushi.

The catch position is calculated from the sushi's
trajectory. We only move the patrons along the plane,
never jump them in the air. This means that sometimes
the patrons must move backwards or forwards so that
the sushi doesn't soar over them.

Probably makes the game too easy at the moment.
We'll make later changes to slow down the top
speed.

Tested: Works on Linux.
Change-Id: I536e55dd7ddf5daf02ff8884329ec37555cda71d