Orient the croc and bird to face forward at angle 0.
Previously the croc and bird would face mostly backwards.
This was awkward because the new movement logic
assumed that angle 0 for a character means forwards.

This change rotates the croc prototype to have angle 0
being forward, and counter-rotates all the instantiations
accordingly.

Also, move the contact regions to match, and expand the
contact reagion so that the croc can eat from both sides.

Tested: Looks good on Linux.
Change-Id: Id78196af8729b021d99f8d959584302a697e853c