Animate the Hungry Hippo with rigged animations.
Update Hippo model with rigged version.
Add animations that use the rigged hippo model.

I've added a timer to the patron states too, because
some of the animations are cyclic and don't exit
on their own.

Also split the "fed" state into "eating" and "satisfied"
states, since these are two different animations for
the hippo, and for the other patrons too (in the near
future).

Tested: Works on Linux and Mac.
Change-Id: I05c421086bc06c61cff7b6ef5698e7e54e30ab28