Added void inputs and outputs and events.
Added an Attribute module that can get attributes off of Entities.
Currently this is just used to display the number of points you have,
and is output to the console.

Added a function to get the Player entity. This is a temporary solution,
in the long term both the RaftEntityNode and PlayerEntityNode should go
away and be replaced by a more general entity node, but this was quick
and easy and does what I need for now.

Made some minor adjustments to the logical operator nodes so that they
fire void events when they become true and when they become false,
rather than just having a boolean output. This will be useful for
setting up logic like "When the timer is less than X *and* when the
player's score is less than Y, then end the game."

Change-Id: I71dcdd695ea1aad113d50b13b97286ecf1375780