Added the ability to vary the river width and banks per zone.
Added a default layout that has the river get wider and wider as you
get towards the end of the lap, then back to the default width.

Added a riverbank override for the middle part of the level to
visibly flatten the riverbank - just as an example.

Change-Id: I63e63209e7189121697e20415ef756abdd2b23d0
Tested: On Linux.