Fixed crash on editing PatronDef.
When ActionDef's name changed, we needed to change its name in the
places we copy it via CopyTable.

This only affected cases when the ActionDef was in memory that the
entity factory was not going to keep around, so it needed to copy
the data out via CopyTable.

Change-Id: I2f25664435b91e4901c3863c250091a7904a7094
Tested: In editor on Linux