Adding more options for patrons catching sushi.
Adding a max angle the patron is willing to turn away from the raft,
to prevent eating sushi while facing away. Making the distance they are
willing to move to catch sushi adjustable in the definition, for more
control in the future.

Also reducing the max time considered, as we don't want patrons taking
too long should they fail for some reason.

Tested on Linux.

Change-Id: Id631f14fcc62311000d28d67d4c4664dd15ab0f0