Fix for Android building from scratch every time.
The issue was caused by the dependency the source files had on the
generated files. Because of this, checking the generated files would
make it as newer, thus forcing all the source files to need to be
rebuilt.

This can be prevented with order-only prerequisites, which lets the
files depend on the generated files being checked first, but not have
the act of checking them be enough to consider them newer.

Tested on Android.

Change-Id: I75111a9177d89d47fce24b3b1976d4a6c40622a5