Allow maximum texture size to be specified per-asset.
Needed for the new unified skybox texture. It is a little
blurry at 1k x 1k.

To use, add an entry to "mesh_meta" in asset_meta.json,
with the "texture_size" specified. This is the maximum
texture size used for all assets that have "name" in
their name. For example,
    { "name": "Skybox", "texture_size": 2048 },

Change-Id: I908a8a0fcb09522192f81475d243923d23949a65
Tested: works on skybox textures.