Updated modules to use new node interface.
The interface in the event library changed, so various changes were
needed in Zooshi to accomodate it. This check in updates the code to be
able to use the new event system (but does not actually use it yet). In
a follow up commit the new graphs will be added and the old event system
will be torn out completely.

Tested on Linux.

Change-Id: I48a29ca5b5e3263c95832bb3c933843c1ee31871