Added a node to change state, used it to end game.
Added a new Graph node that can be used to end the game and bring you
back to the title screen. Each lap now requires that you get 15 more
points than the previous lap.

Fixed a bug that was causing the new_lap broadcaster to not be hit.
There was some obsolete code that was gating it, so that code was
removed.

Tested on Linux.

Change-Id: I844e279839748bfca2b032b89d46621c83272c4a