Fix hang in OSX build.
On the update thread, we called Input::AdvanceFrame(), which then calls
SDL_PollEvent() in order to processes OS messages.

However, from the SDL documentation on SDL_PollEvent():
(https://wiki.libsdl.org/SDL_PollEvent)
"As this function implicitly calls SDL_PumpEvents(), you can only call
this function in the thread that set the video mode."

So, we have to move the Input::AdvanceFrame() call to the render thread,
unfortunately.

Change-Id: Ia2c0412d561d752cb006f3d04afe1f3a6df9db23
Tested: Works on Mac.