Seperating how the projectile offset is defined.
Breaking apart how the projectile offset was defined into height and
forward amounts. Height is along the world's Z-axis, while forward is
relative to the direction that the sushi is being thrown. This allows
the sushi to spawn farther away from the camera, preventing near
clipping issues.

Tested on Linux and Android.

Change-Id: Ic06874d94a1882b456511c107fae8889d74f9833