Get Visual Studio 2012 running.
I've added a hack to get the code to run on VS2012,
but the existing layout is fragile. It breaks depending on
if the implementation of std::vector::push_back uses
in-place creation (like Clang) or copy creation (like VS).

Also note that Clang seems to choose the intrinsically
generate the move constructor first. VS intrinsically generates
the copy constructor first. Regardless of which is compliant,
it would be nice if our code supported both without resorting
to hacks such as the one in this change.

Tested: Works on Linux and Windows.
Change-Id: Ib5384c64c858e6b5ac01bcdfa72a99c036ce9dfd