Pause audio when entering pause state.
When transitioning between states, the OnEnter and OnExit functions are
now aware of what states they're transitioning between. This is used to
tell whether to pause or outright stop the music when leaving the
gameplay state (music is paused on the pause screen, it's stopped if
returning to the menu screen).

Also fixed a bug where the audio would not pause when the game lost
focus.

Bug: 23089533, 23969015
Change-Id: I746a0c0ced3eeb0d33a66cd5707111313ccff8a5