Moved the GraphComponent to the component library.
This required making a lot of minor changes to references to the graph
component.

Also added an Animation module with an AnimationCompleteNode for firing
events when an animation completes. Also updated how CollisionData
events work, and updated the graphs to line up with the new scheme.

Tested on Linux, Mac and Android.

Change-Id: I473ba9f60bc56130afea3bf1e6e15ccab2b779ad