Fixing how rail denizens inherited transforms.
How rail denizen data inherited transforms had issues with the editor,
as it was accumulating the tranform data into the offset data, so every
reload from the editor essentially doubled the position the offset was
using. Fixed this by splitting the offset and the transform data,
and undoing the rail changes when entering the editor.

Tested on Linux, iterating with the editor.

Change-Id: I8ece1356bcacf2f7c1fdec294ce50eec4d44e2f2