Update introbox model.
Used to be created from a json file, now is an fbx file.

Increased scale slightly. The box geo was getting culled
when debugging on desktop (the introbox doesn't normally
appear on desktop since it's a Cardboard feature). It's
nice to be able to see it on desktop, however.

Tested: Works on Mac and Android.
Change-Id: Iaad156fb774eaa647db3d0b86df7380f57f70030