Ensure sushi thrown when introbox button pressed.
There were four problems:
(1) RenderPrep() was not implemented,
(2) the sushi physics was pulling the sushi to the top of the screen,
(3) the box was too small, and
(4) the fade out happened too quickly.

For (2), I brought the box up to +30 height, instead of -30.
For (3), I increased the size of the introbox x5.
For (4), I added a short timer after the sushi's thrown and
before the fade-out begins.

Change-Id: Ie9905b4debaddf00da4d8c77d269d72162ed4956
Tested: Works on Android. Also tested Mac works normally too.