Added target score to scoreboard.
The scoreboard now displays both the current score and the target score.

The score board has been moved to the right side. Since we're moving
clockwise around the course, you're more likely to be looking to the
right than to the left, so this makes it easier to see.

Added a 100's digit to the scoreboard, since getting a score that high
isn't that unreasonable.

Due to how we sort meshes for the alpha pass, I had to put in a hack to
get the digits to render correctly. Meshes are sorted based on the
distance from the camera to the entity, so I moved the digit entities
forward and made the digit meshes pushed back slightly. This ensures
the digits do not clip the sign they're supposed to be on.

Tested on Linux and Android.

Change-Id: Ia7624a01d70d900975180a79e9bc09ab77765451