Removed asyncronous callbacks on shutdown.
On desktop these don't cause a problem, but due to how android works
these asyncronous callbacks might be called after the game is supposed
to have ended, so it's important to clear them up or the game will crash
on shutdown.

Also updated the manifest to include the necessary permissions.

Tested on Linux and Android.

Change-Id: Id5f97ece1897b83ad1a810b20ef0d7b897948d53