Updated scoring system; added new SFX.
Your target score is now 20 + your current score instead of 15 + your
target score. That way you can't bank a ton of points for each new lap.

I also noticed there were some sound effects that never made it into the
game, so I added them in, and also hooked up some of the sounds as a
victory and failure chime that plays when a lap is complete to make it
more obvious that something happened when you finish a lap.

Tested on Linux and Android.
Bug: 24702072
Change-Id: I68ea270578f5cd939ec459a73b7fc2eea35b86f3