Improve sushi trajectory.
Sushi's initial position was tied to the player's position,
but the axes were swapped. This meant that when you looked up,
the sushi would be launched upside down--something no decent
sushi chef would do. Now the sushi is always launched upright.

The sushi rotation was also always similar: it would always
rotate with backspin, clockwise, and forward. Now the rotation
direction is randomized.

We rotate less in the pitch and yaw directions now. This keeps
the sushi more upright for longer, as you would hope for in a
sushi restaurant.

Change-Id: Ia3ad778e730eefe29963544731de4f558cd7861e
Tested: Works well on Mac.