Reducing uniform usage in shader bone_transforms.
Since the 'w' row of the mat4 for 'bone_transforms' is always a
constant (0, 0, 0, 1), it is not essential for the 'bone_transforms'
array. Therefore, we can use an array of vec4's that contain only
the essential information, and then use a helper function to add
in the constant (0, 0, 0, 1) as the 'w' row whenever a complete
mat4 is required.

Bug: b/23494891
Change-Id: I2b7cfb23916a329b2273c64be2f625e8a90b8351
Tested: This was tested by compiling and running Zooshi on OS X.