Patrons appear and wave at end of game.
Fixes problem where we'd kick out of Cardboard
as soon as the game ended--a jarring experience.

Now the game stays in Cardboard. It just
starts a new game after the current one ends
when the button is pressed

Also attempts to draw attention to the score
sign at the end of the game, since the bird is
the last to disappear and is right on top of the
sign.

Added concept of events to the patrons.
When an event is triggered, the patrons will execute
a series of animations.

Tested: Looks good on Linux and Cardboard.
Change-Id: Ie4bab2f9437360b5dc08f0f99a8f15841b70b138