Onscreen controller
Set ZOOSHI_FORCE_ONSCREEN_CONTROLLER to 1 in game.cpp to force on desktop
or mobile builds.  The onscreen controller *should* (not tested) be enabled
if the device doesn't have a gyro and hence doesn't support Cardboard.

Clamped yaw when using a gamepad on the onscreen controller to avoid
gimbal lock.

Also, ran clang-format on each file so there are a few formatting changes.

Tested:
Verified the on-screen controller works on Android and desktop.
Verified the on-screen controller is enabled when
SupportsHeadMountedDisplay() returns false and disabled otherwise.
Verified the mouse controller still works on the desktop by default.

Bug: 24953569
Change-Id: I57c3ecc5b2d2f7e7ceb400fb06080c9dac565a52