Forcing the screen to stay on with Android.
Adding the flag on start up that keeps the device from automatically
turning off the screen while the game is running. This is especially
important for Cardboard mode, as the screen may not be getting hit
enough to stay on normally. Note that this does not affect turning off
the screen manually.

Tested on Android.

Change-Id: I9c6c4498515b7c764d72ade7acb38a59177d3fd9