Forcing the cameras to update on changing state.
Each state maintains its own camera, which is normally updated in the
update loop. However, the first frame is rendered before the state has
an update, so that first frame might be rendered from an incorrect
camera. Adding calls on the states to update the camera on enter, to
prevent this.

Tested on Linux and Android.

Change-Id: I263f3d98602676714d1894c1d529e7bf2a4c3820