Reseting the facing when exiting to the menu.
When returning to the menu from the game the controller was not being
reset, so it would end up facing the same direction from when the game
was left, instead of towards the starting hippo. Adding a reset to the
controller facing to address that.

Tested on Linux and Android.

Change-Id: Ic2aa8930ce88632433ed430c56235a4dbb021fe4