Fixing generated file dependencies on schemas.
The files generated from schemas were set to depend on the motive and
pindrop libraries, instead of just the schema files in them. This
caused the files to be regenerated with any change to those libraries,
thus causing most of Zooshi to rebuild. Fixed to only depend on the
schema files.

Tested on Linux, editing code in motive did not force generation,
while editing motive schema file still does.

Change-Id: I1a2e9155eb31181ab44d934132aae61a34287256