Added a user_tag to the river, and a sushi graph.
A sushi graph has been added which triggers a sound when it collides
with the ground. In order to properly detect collisions with the river
and riverbanks, they needed to support having a user tag like patrons
have, so that was included as well and set to "ground".

Tested on Linux.

Change-Id: I2c95b16f6fe334070f33adc5a853600879d4def8