Updated audio and made sushi disappear in water.
Added audio for hitting the water and a few other missed things. When
Sushi hits the water, it now disappears. Moved the ground plane down
half a meter so that the water submerges before disappearing.

Removed the code that was taking away mouse control on the loading
screen.

Tested on Android and Linux.

Change-Id: Ib31d3b3316208e0528cbe865328d38c08abe1c2d