Organized module initialization.
Breadboard now has each module handle the registration of any special
data types they need to use, so it's no longer necessary to register
types in the game initialization. Also added using directives so that
breadboard:: doesn't have to be prefixed on everything in the modules
cpp files.

Tested on Linux and Android.

Change-Id: Ic18f4df6841474af6f88523e0faeb7c5d969862d