Calm down patron motion, especially bird and croc.
Those two were searching for new targets with very high
frequency. It caused their movement to be erratic.

I added a new variable to control how often a patron searches
for a new target.

I also damped the motion down a bit, especially for the croc
rotation and the bird search distance.

Change-Id: I54401048bd5779c35fcfdb295961e7628c5ddccf
Tested: Works on Linux and Android.