Ensure meshes have the correct scale.
The new --unit option in the mesh_pipeline ensures that,
no matter how the meshes are export to the FBX, they're brought
in at the proper scale.

Note that some of the assets are authored in m instead of cm,
so we need to pass them in a new unit.

Tested: Assets come through pipeline exactly as before,
even if they're authored 100x bigger than others.

Change-Id: I91d2b41c6749971a886c1891576429b55bad9fb1