Fixed a crash in Play Games login.
When fpl_base was renamed to fplbase Google Play Login broke. The game would
crash becuse Java expected a certain naming convention for C++ functions
and it could no longer find the function it needed to call into because
it hadn't been renamed on the C++ side. This rename solves that issue.

Tested on Android.

Change-Id: Ib38f1261e080adcad10225d1ed5d4ef1e4b1968a