Fix for the new lap event.
Sometimes new lap events would not fire due to how the math was being
done. The math for the spline has been simplified an no longer makes a
call to floor(). The issue was that if the spline progress percentage
was exactly 1.0, it would get added to floor(previous_percent). Because
this value is larger than the previously seen value, it would not call
the new lap event.

I've tested this on Android and Linux and it definitely doesn't make
anything worse, but because this bug is so inconsistent it's hard to
verify that it's no longer happening.

Change-Id: Ibb78fa569ca3930ebf68f884ddf5b6e470d7ab81