Simplified Graphs and Fixed a few errors.
Removed the *DataRef (e.g. TransformDataRef) edge types. They didn't
actually save us any memory or speed things up at all, and they only
made the graphs more complicated. Things that used to take a *DataRef
now just take entities instead.

A lot of graphs had redundant nodes or nodes responding to events they
shouldn't, and this would cause errors like incrementing the target
score sometimes when it shouldn't. These have been cleaned up.

Also changed the sound effects on sushi to only play when the speed of
the collision exceeds an arbitrary threshold so that we don't get a
machine gun sound any time they land on the ground. It still happens to
an extent when sushi bounces around a lot rapidly, but it's not nearly
as bad as before.

Tested on Android and Linux.
Bug: 25279644
Change-Id: Id4a5ca1245b1d9ab5837ab56ea1bb9aefa84f533